+++ nes-benchmark/src/MicroBenchmarks/BenchmarkOutputChannel.cpp
@@ -179,9 +179,9 @@ int main(int argc, char** argv) {
     NES::Benchmark::Util::createRangeVector<uint64_t>(allSenderThreads, 4, 8, 2);
     NES::Benchmark::Util::createRangeVector<uint64_t>(allServerThreads, 4, 8, 2);
     NES::Benchmark::Util::createRangeVector<uint64_t>(allDataSizesToBeSent,
-                                                      (uint64_t) 1 * 1024 * 1024 * 1024,
-                                                      (uint64_t) 4 * 1024 * 1024 * 1024,
-                                                      (uint64_t) 1 * 1024 * 1024 * 1024);
+                                                      1_u64 * 1024 * 1024 * 1024,
+                                                      4_u64 * 1024 * 1024 * 1024,
+                                                      1_u64 * 1024 * 1024 * 1024);
 
     std::string benchmarkName = "DataChannel";
     std::string benchmarkFolderName = "DataChannel";
