+++ nes-runtime/src/Execution/Operators/Relational/Aggregation/BatchAggregation.cpp
@@ -44,7 +44,7 @@ BatchAggregation::BatchAggregation(
 
 void BatchAggregation::setup(ExecutionContext& ctx) const {
     auto globalOperatorHandler = ctx.getGlobalOperatorHandler(operatorHandlerIndex);
-    Value<UInt64> entrySize = (uint64_t) 0;
+    Value<UInt64> entrySize = 0_u64;
     for (auto& function : aggregationFunctions) {
         entrySize = entrySize + function->getSize();
     }
