+++ nes-runtime/tests/UnitTests/Execution/Expressions/TextFunctions/SimilarityFunctions/LevenshteinTest.cpp
@@ -51,29 +51,29 @@ TEST_F(LevenshteinTest, BaseTest) {
     auto textValue0 = Value<Text>("db");
     auto textValue1 = Value<Text>("duck");
     auto dist1 = expression.eval(textValue0, textValue1);
-    EXPECT_EQ(dist1, (uint64_t) 3);
+    EXPECT_EQ(dist1, 3_u64);
 
     auto textValue2 = Value<Text>("aro");
     auto dist2 = expression.eval(textValue2, Value<Text>("roa"));
-    EXPECT_EQ(dist2, (uint64_t) 2);
+    EXPECT_EQ(dist2, 2_u64);
 
     auto dist3 = expression.eval(textValue2, Value<Text>("oar"));
-    EXPECT_EQ(dist3, (uint64_t) 2);
+    EXPECT_EQ(dist3, 2_u64);
 
     auto dist4 = expression.eval(textValue2, Value<Text>("daro"));
-    EXPECT_EQ(dist4, (uint64_t) 1);
+    EXPECT_EQ(dist4, 1_u64);
 
     auto dist5 = expression.eval(textValue2, Value<Text>("ao"));
-    EXPECT_EQ(dist5, (uint64_t) 1);
+    EXPECT_EQ(dist5, 1_u64);
 
     auto textValue3 = Value<Text>("duck");
     auto dist6 = expression.eval(textValue1, textValue3);
-    EXPECT_EQ(dist6, (uint64_t) 0);
+    EXPECT_EQ(dist6, 0_u64);
 
     auto textValue4 = Value<Text>("d");
     auto textValue5 = Value<Text>("z");
     auto dist7 = expression.eval(textValue4, textValue5);
-    EXPECT_EQ(dist7, (uint64_t) 1);
+    EXPECT_EQ(dist7, 1_u64);
 }
 
 TEST_F(LevenshteinTest, FailTest) {
